refactor: remove unused backon dependency and retry logic from orch.rs; enhance retry handling in provider.rs
