+++ crates/forge_services/src/provider.rs
@@ -1,8 +1,12 @@
 use std::sync::Arc;
 
 use anyhow::{Context, Result};
-use forge_domain::{ChatCompletionMessage, Context as ChatContext, Model, ModelId, ResultStream};
+use backon::{ExponentialBuilder, Retryable};
+use forge_domain::{
+    ChatCompletionMessage, Context as ChatContext, Error, Model, ModelId, ResultStream, RetryConfig,
+};
 use forge_provider::Client;
+use tracing::warn;
 
 use crate::services::{EnvironmentService, ProviderService};
 use crate::Infrastructure;
@@ -11,6 +15,7 @@ use crate::Infrastructure;
 pub struct ForgeProviderService {
     // The provider service implementation
     client: Arc<Client>,
+    retry_config: RetryConfig,
 }
 
 impl ForgeProviderService {
@@ -18,9 +23,11 @@ impl ForgeProviderService {
         let infra = infra.clone();
         let env = infra.environment_service().get_environment();
         let provider = env.provider.clone();
+        let retry_config = env.retry_config.clone();
         let version = env.version();
         Self {
-            client: Arc::new(Client::new(provider, env.retry_config, version).unwrap()),
+            client: Arc::new(Client::new(provider, retry_config.clone(), version).unwrap()),
+            retry_config,
         }
     }
 }
@@ -32,8 +39,15 @@ impl ProviderService for ForgeProviderService {
         model: &ModelId,
         request: ChatContext,
     ) -> ResultStream<ChatCompletionMessage, anyhow::Error> {
-        self.client
-            .chat(model, request)
+        let retry_config = &self.retry_config;
+        (|| self.client.chat(model, request.clone()))
+            .retry(
+                ExponentialBuilder::default()
+                    .with_factor(retry_config.backoff_factor as f32)
+                    .with_max_times(retry_config.max_retry_attempts)
+                    .with_jitter(),
+            )
+            .when(should_retry)
             .await
             .with_context(|| format!("Failed to chat with model: {model}"))
     }
@@ -45,4 +59,13 @@ impl ProviderService for ForgeProviderService {
     async fn model(&self, model: &ModelId) -> Result<Option<Model>> {
         self.client.model(model).await
     }
+}
+
+fn should_retry(error: &anyhow::Error) -> bool {
+    let retry = error
+        .downcast_ref::<Error>()
+        .is_some_and(|error| matches!(error, Error::Retryable(_, _)));
+
+    warn!(error = %error, retry = retry, "Retrying on error");
+    retry
 }
